+++ rust/src/datatypes/schema.rs
@@ -1,0 +1,412 @@
+// Copyright 2023 Lance Developers.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Schema
+
+use std::{
+    collections::HashMap,
+    fmt::{self, Debug, Formatter},
+};
+
+use arrow_array::RecordBatch;
+use arrow_schema::{Field as ArrowField, Schema as ArrowSchema};
+
+use super::field::Field;
+use crate::arrow::*;
+use crate::{format::pb, io::object_reader::ObjectReader, Error, Result};
+
+/// Lance Schema.
+#[derive(Default, Debug, PartialEq, Clone)]
+pub struct Schema {
+    /// Top-level fields in the dataset.
+    pub fields: Vec<Field>,
+    /// Metadata of the schema
+    pub metadata: HashMap<String, String>,
+}
+
+impl Schema {
+    /// Project the columns over the schema.
+    ///
+    /// ```ignore
+    /// let schema = Schema::from(...);
+    /// let projected = schema.project(&["col1", "col2.sub_col3.field4"])?;
+    /// ```
+    pub fn project<T: AsRef<str>>(&self, columns: &[T]) -> Result<Self> {
+        let mut candidates: Vec<Field> = vec![];
+        for col in columns {
+            let split = col.as_ref().split('.').collect::<Vec<_>>();
+            let first = split[0];
+            if let Some(field) = self.field(first) {
+                let projected_field = field.project(&split[1..])?;
+                if let Some(candidate_field) = candidates.iter_mut().find(|f| f.name == first) {
+                    candidate_field.merge(&projected_field)?;
+                } else {
+                    candidates.push(projected_field)
+                }
+            } else {
+                return Err(Error::Schema(format!(
+                    "Column {} does not exist",
+                    col.as_ref()
+                )));
+            }
+        }
+
+        Ok(Self {
+            fields: candidates,
+            metadata: self.metadata.clone(),
+        })
+    }
+
+    pub fn project_by_ids(&self, column_ids: &[i32]) -> Result<Self> {
+        let protos: Vec<pb::Field> = self.into();
+
+        let filtered_protos: Vec<pb::Field> = protos
+            .iter()
+            .filter(|p| column_ids.contains(&p.id))
+            .cloned()
+            .collect();
+        Ok(Self::from(&filtered_protos))
+    }
+
+    /// Exclude the fields from `other` Schema, and returns a new Schema.
+    pub fn exclude<T: TryInto<Self> + Debug>(&self, schema: T) -> Result<Self> {
+        let other = schema.try_into().map_err(|_| {
+            Error::Schema("The other schema is not compatible with this schema".to_string())
+        })?;
+        let mut fields = vec![];
+        for field in self.fields.iter() {
+            if let Some(other_field) = other.field(&field.name) {
+                if field.data_type().is_struct() {
+                    if let Some(f) = field.exclude(other_field) {
+                        fields.push(f)
+                    }
+                }
+            } else {
+                fields.push(field.clone());
+            }
+        }
+        Ok(Self {
+            fields,
+            metadata: HashMap::default(),
+        })
+    }
+
+    /// Get a field by name. Return `None` if the field does not exist.
+    pub fn field(&self, name: &str) -> Option<&Field> {
+        let split = name.split('.').collect::<Vec<_>>();
+        self.fields
+            .iter()
+            .find(|f| f.name == split[0])
+            .map(|c| c.sub_field(&split[1..]))
+            .flatten()
+    }
+
+    pub(crate) fn field_id(&self, column: &str) -> Result<i32> {
+        self.field(column)
+            .map(|f| f.id)
+            .ok_or_else(|| Error::Schema("Vector column not in schema".to_string()))
+    }
+
+    /// Recursively collect all the field IDs,
+    pub(crate) fn field_ids(&self) -> Vec<i32> {
+        // TODO: make a tree traversal iterator.
+        let protos: Vec<pb::Field> = self.into();
+        protos.iter().map(|f| f.id).collect()
+    }
+
+    pub(crate) fn mut_field_by_id(&mut self, id: i32) -> Option<&mut Field> {
+        for field in self.fields.as_mut_slice() {
+            if field.id == id {
+                return Some(field);
+            }
+            if let Some(grandchild) = field.mut_field_by_id(id) {
+                return Some(grandchild);
+            }
+        }
+        None
+    }
+
+    pub(crate) fn max_field_id(&self) -> Option<i32> {
+        self.fields.iter().map(|f| f.max_id()).max()
+    }
+
+    /// Load dictionary value array from manifest files.
+    pub(crate) async fn load_dictionary<'a>(&mut self, reader: &dyn ObjectReader) -> Result<()> {
+        for field in self.fields.as_mut_slice() {
+            field.load_dictionary(reader).await?;
+        }
+        Ok(())
+    }
+
+    /// Recursively attach set up dictionary values to the dictionary fields.
+    pub(crate) fn set_dictionary(&mut self, batch: &RecordBatch) -> Result<()> {
+        for field in self.fields.as_mut_slice() {
+            let column = batch.column_by_name(&field.name).ok_or_else(|| {
+                Error::Schema(format!(
+                    "column '{}' does not exist in the record batch",
+                    field.name
+                ))
+            })?;
+            field.set_dictionary(column);
+        }
+        Ok(())
+    }
+
+    fn set_field_id(&mut self) {
+        let mut current_id = self.max_field_id().unwrap_or(-1) + 1;
+        self.fields
+            .iter_mut()
+            .for_each(|f| f.set_id(-1, &mut current_id));
+    }
+
+    pub fn merge(&self, other: &Self) -> Self {
+        let mut fields = self.fields.clone();
+        for field in other.fields.as_slice() {
+            if !fields.iter().any(|f| f.name == field.name) {
+                fields.push(field.clone());
+            }
+        }
+        let mut metadata = other.metadata.clone();
+        self.metadata.iter().for_each(|(k, v)| {
+            metadata.insert(k.to_string(), v.to_string());
+        });
+        Self { fields, metadata }
+    }
+}
+
+impl fmt::Display for Schema {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        for field in self.fields.iter() {
+            writeln!(f, "{field}")?
+        }
+        Ok(())
+    }
+}
+
+/// Convert `arrow2::datatype::Schema` to Lance
+impl TryFrom<&ArrowSchema> for Schema {
+    type Error = Error;
+
+    fn try_from(schema: &ArrowSchema) -> Result<Self> {
+        let mut schema = Self {
+            fields: schema
+                .fields
+                .iter()
+                .map(|f| Field::try_from(f.as_ref()))
+                .collect::<Result<_>>()?,
+            metadata: schema.metadata.clone(),
+        };
+        schema.set_field_id();
+
+        Ok(schema)
+    }
+}
+
+/// Convert Lance Schema to Arrow Schema
+impl From<&Schema> for ArrowSchema {
+    fn from(schema: &Schema) -> Self {
+        Self {
+            fields: schema.fields.iter().map(ArrowField::from).collect(),
+            metadata: schema.metadata.clone(),
+        }
+    }
+}
+
+/// Convert Lance Schema to Arrow Schema
+impl From<&Schema> for Schema {
+    fn from(schema: &Schema) -> Self {
+        schema.clone()
+    }
+}
+
+/// Convert list of protobuf `Field` to a Schema.
+impl From<&Vec<pb::Field>> for Schema {
+    fn from(fields: &Vec<pb::Field>) -> Self {
+        let mut schema = Self {
+            fields: vec![],
+            metadata: HashMap::default(),
+        };
+
+        fields.iter().for_each(|f| {
+            if f.parent_id == -1 {
+                schema.fields.push(Field::from(f));
+            } else {
+                let parent = schema.mut_field_by_id(f.parent_id).unwrap();
+                parent.children.push(Field::from(f));
+            }
+        });
+
+        schema
+    }
+}
+
+/// Convert a Schema to a list of protobuf Field.
+impl From<&Schema> for Vec<pb::Field> {
+    fn from(schema: &Schema) -> Self {
+        let mut protos: Self = vec![];
+        schema.fields.iter().for_each(|f| {
+            protos.extend(Self::from(f));
+        });
+        protos
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use arrow_schema::{
+        DataType, Field as ArrowField, Fields as ArrowFields, Schema as ArrowSchema,
+    };
+
+    #[test]
+    fn test_schema_projection() {
+        let arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new("a", DataType::Int32, false),
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![
+                    ArrowField::new("f1", DataType::Utf8, true),
+                    ArrowField::new("f2", DataType::Boolean, false),
+                    ArrowField::new("f3", DataType::Float32, false),
+                ])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        let schema = Schema::try_from(&arrow_schema).unwrap();
+        let projected = schema.project(&["b.f1", "b.f3", "c"]).unwrap();
+
+        let expected_arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![
+                    ArrowField::new("f1", DataType::Utf8, true),
+                    ArrowField::new("f3", DataType::Float32, false),
+                ])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        assert_eq!(ArrowSchema::from(&projected), expected_arrow_schema);
+    }
+
+    #[test]
+    fn test_schema_project_by_ids() {
+        let arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new("a", DataType::Int32, false),
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![
+                    ArrowField::new("f1", DataType::Utf8, true),
+                    ArrowField::new("f2", DataType::Boolean, false),
+                    ArrowField::new("f3", DataType::Float32, false),
+                ])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        let schema = Schema::try_from(&arrow_schema).unwrap();
+        let projected = schema.project_by_ids(&[1, 2, 4, 5]).unwrap();
+
+        let expected_arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![
+                    ArrowField::new("f1", DataType::Utf8, true),
+                    ArrowField::new("f3", DataType::Float32, false),
+                ])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        assert_eq!(ArrowSchema::from(&projected), expected_arrow_schema);
+    }
+
+    #[test]
+    fn test_schema_set_ids() {
+        let arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new("a", DataType::Int32, false),
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![
+                    ArrowField::new("f1", DataType::Utf8, true),
+                    ArrowField::new("f2", DataType::Boolean, false),
+                    ArrowField::new("f3", DataType::Float32, false),
+                ])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        let schema = Schema::try_from(&arrow_schema).unwrap();
+
+        let protos: Vec<pb::Field> = (&schema).into();
+        assert_eq!(
+            protos.iter().map(|p| p.id).collect::<Vec<_>>(),
+            (0..6).collect::<Vec<_>>()
+        );
+    }
+
+    #[test]
+    fn test_get_nested_field() {
+        let arrow_schema = ArrowSchema::new(vec![ArrowField::new(
+            "b",
+            DataType::Struct(ArrowFields::from(vec![
+                ArrowField::new("f1", DataType::Utf8, true),
+                ArrowField::new("f2", DataType::Boolean, false),
+                ArrowField::new("f3", DataType::Float32, false),
+            ])),
+            true,
+        )]);
+        let schema = Schema::try_from(&arrow_schema).unwrap();
+
+        let field = schema.field("b.f2").unwrap();
+        assert_eq!(field.data_type(), DataType::Boolean);
+    }
+
+    #[test]
+    fn test_exclude_fields() {
+        let arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new("a", DataType::Int32, false),
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![
+                    ArrowField::new("f1", DataType::Utf8, true),
+                    ArrowField::new("f2", DataType::Boolean, false),
+                    ArrowField::new("f3", DataType::Float32, false),
+                ])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        let schema = Schema::try_from(&arrow_schema).unwrap();
+
+        let projection = schema.project(&["a", "b.f2", "b.f3"]).unwrap();
+        let excluded = schema.exclude(&projection).unwrap();
+
+        let expected_arrow_schema = ArrowSchema::new(vec![
+            ArrowField::new(
+                "b",
+                DataType::Struct(ArrowFields::from(vec![ArrowField::new(
+                    "f1",
+                    DataType::Utf8,
+                    true,
+                )])),
+                true,
+            ),
+            ArrowField::new("c", DataType::Float64, false),
+        ]);
+        assert_eq!(ArrowSchema::from(&excluded), expected_arrow_schema);
+    }
+}
